Introduce LargeStructureUtils class
